Fix lists with blank lines interspersed.
